refactor: move context files from hooks to context folder and update import paths
