+++ spayc-frontend/src/Components/EditPanel/EditPanel.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useRef } from 'react';
 import MappingServices from '../Services/MapServices/MappingServices';
-import { useServicesContext } from '../../hooks/useServicesContext';
+import { useServicesContext } from '../../context/useServicesContext';
 import ServiceCreateForm from './Forms/ServiceCreateForm';
 import { SubmitHandler } from 'react-hook-form';
 import { ServicesType } from '../../utils/Interface';
