+++ spayc-frontend/src/Components/EditPanel/Forms/ServiceCreateForm.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { ServicesType } from '../../../utils/Interface';
-import { useServicesContext } from '../../../hooks/useServicesContext';
+import { useServicesContext } from '../../../context/useServicesContext';
 
 interface ServiceCreateFormProps {
   onSubmit: SubmitHandler<{ service: ServicesType; panelSource: boolean; descriptionArray: string[]; file: FileList | null }>;
