+++ spayc-frontend/src/context/useAuthContext.ts
